python

+++ Actividades bucles/actividad_bucles37.py
@@ -2,7 +2,7 @@
 #o suspendido.
 var_repeticiones=int(input("Introduce el número de notas que quieres introducir:"))
 for i in range(var_repeticiones):
-    var_nota=int(input("Introduce una nota:"))
+    var_nota=float(input("Introduce una nota:"))
     if var_nota>=5:
         print("Asignatura aprobada")
     else:
